organizing process
