+++ server/index.js
@@ -1,0 +1,4 @@
+const { server } = require('./ssr-server');
+const { PORT } = require('./config/config.json');
+
+server.listen(PORT, () => console.log('Server listening on %s', PORT));
